Makes astar benchmark run.
valid_exit() in the benchmark will receive negative x and y deltas
from get_available_exist() which when the point they are examining a loc
which is x=0 or y=0 would create a negative search position.  This in
turn would panic in algorithm2d when trying to coerce them into usize.