Scaler2 (#246)
Completely changes the approach to rendering with gutters and nice scaling.
The framebuffer is now always the natural size of the target. No scaling artifacts, and it looks good. FB only resizes when you are using auto-scaling consoles.
Framebuffer to screen uses linear filtering for a nicer scaling experience.
Should assist with: #226 #220 #213 #178 #171

* #220 - Change the framebuffer texture scaling parameters. No artefacts when scaling on my system. Fullscreen/maximize still needs some love.

* Backing buffer is no longer optional for OpenGL mode.

* GL framebuffers now retain a *natural* size at all times, and the framebuffer->screen is scaled (and has gutters applied). No more nasty resize artifacts, slightly nicer linear scaling of the console.

* Repeat scaling changes for webgpu.